+++ provider/helper.go
@@ -9,25 +9,6 @@ import (
 	neon "github.com/kislerdm/neon-sdk-go"
 )
 
-func pgSettingsToMap(v neon.PgSettingsData) map[string]interface{} {
-	o := make(map[string]interface{}, len(v))
-	for k, v := range v {
-		o[k] = v
-	}
-	return o
-}
-
-func mapToPgSettings(v map[string]interface{}) *neon.PgSettingsData {
-	if len(v) == 0 {
-		return nil
-	}
-	o := make(neon.PgSettingsData, len(v))
-	for k, v := range v {
-		o[k] = v
-	}
-	return &o
-}
-
 func intValidationNotNegative(v interface{}, s string) (warn []string, errs []error) {
 	if vv, ok := v.(int); ok && vv < 0 {
 		errs = append(errs, errors.New(s+" must be not negative"))