+++ provider/resource_jwks_url.go
@@ -67,7 +67,7 @@ func resourceJwksUrl() *schema.Resource {
 	}
 }
 
-func updateStateJwksUrl(d *schema.ResourceData, v neon.JWKS, roleNames []string) error {
+func updateStateJwksUrl(d *schema.ResourceData, v neon.JWKS, roleNames *[]string) error {
 	if err := d.Set("project_id", v.ProjectID); err != nil {
 		return err
 	}
@@ -109,7 +109,7 @@ func resourceJwksUrlCreate(ctx context.Context, d *schema.ResourceData, meta int
 		for i, roleName := range vv {
 			roleNames[i] = roleName.(string)
 		}
-		cfg.RoleNames = roleNames
+		cfg.RoleNames = &roleNames
 	}
 	if v, ok := d.GetOk("branch_id"); ok && v.(string) != "" {
 		cfg.BranchID = pointer(v.(string))
@@ -168,7 +168,7 @@ func resourceJwksUrlDelete(ctx context.Context, d *schema.ResourceData, meta int
 	client := meta.(*neon.Client)
 	resp, err := client.DeleteProjectJWKS(d.Get("project_id").(string), d.Id())
 	if err == nil {
-		err = updateStateJwksUrl(d, resp, []string{})
+		err = updateStateJwksUrl(d, resp, nil)
 	}
 	if err == nil {
 		d.SetId(resp.ID)