+++ provider/acc_test.go
@@ -1294,3 +1294,114 @@ func testQuery(uri string) error {
 
 	return err
 }
+
+func TestAccMaintenanceWindow(t *testing.T) {
+	if os.Getenv("TF_ACC") != "1" {
+		t.Skip("TF_ACC must be set to 1")
+	}
+
+	client, err := neon.NewClient(neon.Config{Key: os.Getenv("NEON_API_KEY")})
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	projectNamePrefix := projectNamePrefix + "-maintenanceWindow-"
+	t.Cleanup(func() {
+		resp, _ := client.ListProjects(nil, nil, &projectNamePrefix, nil, nil)
+		for _, project := range resp.Projects {
+			_, _ = client.DeleteProject(project.ID)
+		}
+	})
+
+	projectName := projectNamePrefix + strconv.FormatInt(time.Now().UnixMilli(), 10)
+
+	var (
+		wantWeekdays  = []int{1, 2}
+		wantStartTime = "21:00"
+		wantEndTime   = "23:00"
+	)
+
+	var weekdaysStr string
+	for i, day := range wantWeekdays {
+		weekdaysStr = weekdaysStr + strconv.Itoa(day)
+		if i < len(wantWeekdays)-1 {
+			weekdaysStr += ","
+		}
+	}
+
+	resource.Test(
+		t, resource.TestCase{
+			ProviderFactories: map[string]func() (*schema.Provider, error){
+				"neon": func() (*schema.Provider, error) {
+					return newAccTest(), nil
+				},
+			},
+			Steps: []resource.TestStep{
+				{
+					ResourceName: "neon_project.this",
+					Config: fmt.Sprintf(`resource "neon_project" "this" { 
+	name = "%s"
+	maintenance_window {
+		weekdays   = [%s]
+	    start_time = "%s"
+	    end_time   = "%s"
+	}
+}`, projectName, weekdaysStr, wantStartTime, wantEndTime),
+					Check: resource.ComposeTestCheckFunc(
+						resource.TestCheckResourceAttr(
+							"neon_project.this", "maintenance_window.#", "1",
+						),
+						resource.TestCheckResourceAttr(
+							"neon_project.this", "maintenance_window.0.start_time", wantStartTime,
+						),
+						resource.TestCheckResourceAttr(
+							"neon_project.this", "maintenance_window.0.end_time", wantEndTime,
+						),
+						resource.TestCheckResourceAttr(
+							"neon_project.this", "maintenance_window.0.weekdays.#",
+							strconv.Itoa(len(wantWeekdays)),
+						),
+
+						func(state *terraform.State) error {
+							ref, err := readProjectInfo(client, projectName)
+							if err != nil {
+								return err
+							}
+
+							if !assert.ElementsMatch(t, ref.Settings.MaintenanceWindow.Weekdays, wantWeekdays) {
+								return fmt.Errorf("the set maintenance weekdays got: %v, want: %v",
+									ref.Settings.MaintenanceWindow.Weekdays, wantWeekdays)
+							}
+
+							if ref.Settings.MaintenanceWindow.StartTime != wantStartTime {
+								return fmt.Errorf("the set maintenance weekdays")
+							}
+
+							return nil
+						},
+					),
+				},
+				{
+					ResourceName: "neon_project.this",
+					Config:       fmt.Sprintf(`resource "neon_project" "this" {name = %s}`, projectName),
+					ImportState:  true,
+					Check: resource.ComposeTestCheckFunc(
+						resource.TestCheckResourceAttr(
+							"neon_project.this", "maintenance_window.#", "1",
+						),
+						resource.TestCheckResourceAttr(
+							"neon_project.this", "maintenance_window.0.start_time", wantStartTime,
+						),
+						resource.TestCheckResourceAttr(
+							"neon_project.this", "maintenance_window.0.end_time", wantEndTime,
+						),
+						resource.TestCheckResourceAttr(
+							"neon_project.this", "maintenance_window.0.weekdays.#",
+							strconv.Itoa(len(wantWeekdays)),
+						),
+					),
+				},
+			},
+		},
+	)
+}