+++ CHANGELOG.md
@@ -11,6 +11,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added the resource `neon_org_api_key` to manage API keys on per org. level.
 
+### Changed
+
+- Updated dependencies:
+  - Neon Go SDK: [v0.16.0](https://github.com/kislerdm/neon-sdk-go/compare/v0.13.0...v0.16.0)
+
 ## [v0.10.0] - 2025-10-12
 
 ### Removed