+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.12.0] - 2025-11-04
+
+### Added
+
+- [[#126](https://github.com/kislerdm/terraform-provider-neon/issues/126)] Added the block `maintenance_window` to the `neon_project` resource to configure the [Neon maintenance window](https://neon.com/docs/manage/updates#updates-on-paid-plans).
+
 ## [v0.11.0] - 2025-10-26
 
 ### Added